Try to estimate idle time between moves
The idle time between moves varies depending on how the ball travels.
It wastes a lot of time to just wait the longest duration for every
move, even short ones.

Having a rough estimate based on the distance travelled works, but it
would be nice to more carefully evaluate the path taken to find a more
accurate measure down the line.