Refactor try_move, general improvements
Removes a lot of usage of cloning/borrowing, better to rely on the
compiler copying in most cases and only clone when necessary.

Movement now has three rough phases, where different rules are
applied depending on the current/next position. This makes it easier
to identify when a rule should be applied.

Some areas are still janky, like what happens when a ball lands on a
slope. This doesn't feel like it's easily avoidable though. So long
as it's passing all the test cases I'm fine with it.