Discard solutions with wasted moves
The solver should not recurse when it wastes a move, otherwise it will
not be able to solve levels with a lot of water and walls in a
reasonable/efficient fashion.

Level 07-07 is a prime example of this, and previously the solver would
get stuck trying paths where it kept sending itself straight into the
water.